Making first fake and first timer.

+++ app/src/main/java/com/example/tomek/blue/BlueEnginePower.java
@@ -38,7 +38,7 @@ public class BlueEnginePower extends BluetoothValue {
         this.valueChanged = true;
 
         try{
-            enginePowerValue = Double.parseDouble(s);
+            enginePowerValue = Integer.parseInt(s);
         } catch (NumberFormatException e){
             this.valueChanged = false;
             Log.e(TAG, "Invalid parse from string to double in BlueSolarVoltage: " + e.getMessage());
